Fix how OpenEdXInstance field defaults are set
Now you can create an instance from the CLI using
inst = OpenEdXInstance.objects.create(name="New Instance name", sub_domain="test1.blah")
